# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solarbeam EMS Cloud",
    "version": "1.0.0",
    "description": "API for Solarbeam EMS Cloud",
    "contact": {
      "email": "vojtechpetrasek@gmail.com"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://opensource.org/license/gpl-3-0/"
    }
  },
  "servers": [
    {
      "url": "https://api.solarbeam.vojtechpetrasek.com",
      "description": "Production server"
    },
    {
      "url": "https://dev.api.solarbeam.vojtechpetrasek.com",
      "description": "Development server"
    },
    {
      "url": "http://localhost:6870",
      "description": "Local server"
    }
  ],
...
  "produces": ["application/json"],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jwt",
        "description": "note: non-oauth scopes are not defined at the securityScheme level"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "apiKey"
      }
    },
    "schemas": {
...
      "DevicesResponse": {
        "type": "object",
        "properties": {
          "active_devices": {
            "type": "array",
            "example": [
              {
                "_id": "5f9e9b7b9c6b4e0017b3b3b3",
                "name": "Device name",
                "sys": {
                  "cts": "2021-01-01T00:00:00.000Z",
                  "mts": "2021-01-01T00:00:00.000Z",
                  "rev": 1
                }
              }
            ]
          },
          "archived_devices": { "type": "array", "example": [] },
          "shared_devices": { "type": "array", "example": [] },
          "errors": {
            "type": "array",
            "example": []
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/devices": {
      "get": {
        "summary": "Devices",
        "tags": ["Devices"],
        "description": "Get devices",
        "security": [
          {
            "bearerAuth": ["read:users", "public"]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid token"
                }
              }
            }
          }
        }
      }
    },
...
    "/device/{id}": {
      "get": {
        "tags": ["Devices"],
        "summary": "Get device by id",
        "description": "Get device by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}